feat(expert): Consolidate ExpertQuestionController and ExpertGroupController into unified ExpertController
Unified expert question and group management into a single `ExpertController`. Simplified routing and removed the `ExpertQuestionController` and `ExpertGroupController` files. Updated imports and main application configuration to use `ExpertController`.
