+++ aihub_api/aihub_api/routes/expert_group/__init__.py
@@ -1,4 +1,3 @@
-from .ExpertGroupController import ExpertGroupController
 from .ExpertGroupService import ExpertGroupService
 
-__all__ = ["ExpertGroupController", "ExpertGroupService"]
+__all__ = ["ExpertGroupService"]
